fix(app_control): 恢复基于调平零点的绝对高度同步控制并优化物理参数换算
- 将 PID 控制目标与安防超限检测统一修正为基于调平零点 (min_mount_halls) 的绝对伸出行程 travel_rel，彻底消除频繁启停累积误差与掉帧误报
  - 将 counts_per_mm 升级为 float 浮点数并配合 roundf() 四舍五入，消除导程为 8mm 时的截断舍入偏差
  - 出厂默认参数更新为导程 8mm、速度 750mm/min，默认安装高度 1m 废除硬编码改为公式动态解算
  - 保留 delta_h 位移增量监控变量，并恢复 comm_error 连续 10 帧通信中断独立急停保护

+++ 2026-273-MCU/APP/app_control.h
@@ -54,7 +54,7 @@ typedef struct {
     float max_travel_diff;              // 4 轴当帧最大绝对高度差 (max - min)
     
     // 自动物理换算参数与安防状态
-    uint32_t counts_per_mm;             // 每 mm 霍尔计数值
+    float counts_per_mm;                // 每 mm 霍尔计数值 (支持 112.5f 浮点精度)
     int32_t max_sync_diff_hall;         // 最大同步差霍尔计数值
     int32_t max_travel_hall;            // 最大行程霍尔计数值
     int16_t calc_base_rpm;              // 算出的基准 RPM
